backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx>=0.27
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy>=1.26,<2.0